requirements.txt
streamlit
google-generativeai
yfinance
pandas
python-dotenv
requests
